GCMemcardDirectory: Move GCIFile class to its own file.
