+++ Source/Core/Core/CMakeLists.txt
@@ -107,6 +107,7 @@ add_library(core
   HW/EXI/EXI_DeviceMic.cpp
   HW/GCKeyboard.cpp
   HW/GCKeyboardEmu.cpp
+  HW/GCMemcard/GCIFile.cpp
   HW/GCMemcard/GCMemcard.cpp
   HW/GCMemcard/GCMemcardDirectory.cpp
   HW/GCMemcard/GCMemcardRaw.cpp
